+++ src/color.h
@@ -4,7 +4,7 @@
 #include "common.h"
 
 struct Color {
-	float r, g, b;
+	double r, g, b;
 };
 typedef struct Color fcolor;
 
@@ -18,12 +18,12 @@ typedef struct Color fcolor;
   (macro_dest).g *= (macro_color).g; \
   (macro_dest).b *= (macro_color).b;
 
-static inline fcolor fcolor_new(float r, float g, float b) {
+static inline fcolor fcolor_new(double r, double g, double b) {
 	fcolor c = {r, g, b};
 	return c;
 }
 
-static inline fcolor color_mul(fcolor a, float f) {
+static inline fcolor color_mul(fcolor a, double f) {
 	fcolor c = {};
 	c.r = a.r*f;
 	c.g = a.g*f;
@@ -33,28 +33,28 @@ static inline fcolor color_mul(fcolor a, float f) {
 
 static inline fcolor color_normalize(fcolor a) {
 	// This is kind of silly but it's just easy for now
-	float len = vec3_len(vec3_new(a.r, a.g, a.b));
+	double len = vec3_len(vec3_new(a.r, a.g, a.b));
 	return fcolor_new(a.r / len, a.g / len, a.b / len);
 }
 
 // Returns a new color that is f% color b and the rest color a.
-static inline fcolor color_mix(fcolor a, fcolor b, float f) {
+static inline fcolor color_mix(fcolor a, fcolor b, double f) {
 	fcolor c = {};
-	c.r = (1.0f - f) * a.r + f * b.r;
-	c.g = (1.0f - f) * a.g + f * b.g;
-	c.b = (1.0f - f) * a.b + f * b.b;
+	c.r = (1.0 - f) * a.r + f * b.r;
+	c.g = (1.0 - f) * a.g + f * b.g;
+	c.b = (1.0 - f) * a.b + f * b.b;
 	return c;
 }
 
 // Some macros for common colors
-#define COLOR_RED fcolor_new(1.0f, 0.0f, 0.0f)
-#define COLOR_GREEN fcolor_new(0.0f, 1.0f, 0.0f)
-#define COLOR_BLUE fcolor_new(0.0f, 0.0f, 1.0f)
-#define COLOR_BLACK fcolor_new(0.0f, 0.0f, 0.0f)
-#define COLOR_WHITE fcolor_new(1.0f, 1.0f, 1.0f)
+#define COLOR_RED fcolor_new(1.0, 0.0, 0.0)
+#define COLOR_GREEN fcolor_new(0.0, 1.0, 0.0)
+#define COLOR_BLUE fcolor_new(0.0, 0.0, 1.0)
+#define COLOR_BLACK fcolor_new(0.0, 0.0, 0.0)
+#define COLOR_WHITE fcolor_new(1.0, 1.0, 1.0)
 #define COLOR_VALUE(f) fcolor_new(f, f, f)
 // Color to be used when textures go bad or something
-#define COLOR_UNDEFPURP fcolor_new(1.0f, 0.65f, 1.0f)
+#define COLOR_UNDEFPURP fcolor_new(1.0, 0.65, 1.0)
 
 
 #endif
